+++ packages/polaris-viz/src/components/Labels/utilities/get-widest-label.ts
@@ -1,0 +1,15 @@
+import type {PreparedLabels} from '../types';
+
+export function getWidestLabel(labels: PreparedLabels[]) {
+  let longestIndex = -1;
+  let longestWidth = 0;
+
+  labels.forEach(({truncatedWidth}, index) => {
+    if (truncatedWidth > longestWidth) {
+      longestIndex = index;
+      longestWidth = truncatedWidth;
+    }
+  });
+
+  return labels[longestIndex];
+}