+++ packages/polaris-viz/src/hooks/use-reduced-label-indexes.ts
@@ -1,0 +1,24 @@
+import {useMemo} from 'react';
+
+export interface Props {
+  dataLength: number;
+  skipEveryNthLabel?: number;
+}
+
+export function useReducedLabelIndexes({
+  dataLength,
+  skipEveryNthLabel = -1,
+}: Props) {
+  const visibleIndexesForWideCharts = useMemo(() => {
+    if (dataLength === 0 || skipEveryNthLabel === 1) {
+      return [];
+    }
+
+    return Array(dataLength)
+      .fill(null)
+      .map((_, index) => index)
+      .filter((_, index) => index % skipEveryNthLabel === 0);
+  }, [dataLength, skipEveryNthLabel]);
+
+  return visibleIndexesForWideCharts;
+}