+++ packages/polaris-viz/src/hooks/useHorizontalBarSizes.ts
@@ -1,28 +1,23 @@
 import {useMemo} from 'react';
 import type {Dimensions} from '@shopify/polaris-viz-core';
 
-import {clamp, getTextHeight} from '../utilities';
+import {clamp} from '../utilities';
 import {
-  FONT_SIZE,
   HORIZONTAL_GROUP_LABEL_HEIGHT,
   HORIZONTAL_SPACE_BETWEEN_CHART_AND_AXIS,
   HORIZONTAL_SPACE_BETWEEN_SETS,
   HORIZONTAL_SPACE_BETWEEN_SINGLE,
-  LINE_HEIGHT,
-  MAX_X_AXIS_LINES,
 } from '../constants';
-import type {LabelFormatter} from '../types';
 
 const MIN_BAR_HEIGHT = 6;
 
-interface Props {
+export interface Props {
   chartDimensions: Dimensions;
   isSimple: boolean;
   isStacked: boolean;
-  labelFormatter: LabelFormatter;
+  labelHeight: number;
   seriesLength: number;
   singleBarCount: number;
-  ticks: number[];
 }
 
 // Returns all the data needed to size and position the
@@ -31,35 +26,10 @@ export function useHorizontalBarSizes({
   chartDimensions,
   isSimple,
   isStacked,
-  labelFormatter,
+  labelHeight,
   seriesLength,
   singleBarCount,
-  ticks,
 }: Props) {
-  const bandwidth = chartDimensions.width / ticks.length;
-
-  const tallestXAxisLabel = useMemo(() => {
-    const maxHeight = LINE_HEIGHT * MAX_X_AXIS_LINES;
-
-    const longestLabel = ticks.reduce((prev, cur) => {
-      const label = `${labelFormatter(cur)}`;
-
-      if (label.length > prev.length) {
-        return label;
-      }
-
-      return prev;
-    }, '');
-
-    const height = getTextHeight({
-      text: longestLabel,
-      fontSize: FONT_SIZE,
-      containerWidth: bandwidth / 2,
-    });
-
-    return clamp({amount: height, min: 0, max: maxHeight});
-  }, [ticks, labelFormatter, bandwidth]);
-
   return useMemo(() => {
     const spaceBetweenXAxis = isSimple
       ? 0
@@ -68,7 +38,7 @@ export function useHorizontalBarSizes({
     const simpleHeight = chartDimensions.height + HORIZONTAL_SPACE_BETWEEN_SETS;
 
     const containerHeight = isSimple ? simpleHeight : chartDimensions.height;
-    const xAxisHeight = isSimple ? 0 : tallestXAxisLabel + spaceBetweenXAxis;
+    const xAxisHeight = isSimple ? 0 : labelHeight + spaceBetweenXAxis;
     const chartHeight = containerHeight - xAxisHeight;
 
     const groupHeight = chartHeight / seriesLength;
@@ -93,20 +63,17 @@ export function useHorizontalBarSizes({
     }
 
     return {
-      bandwidth,
       barHeight,
       chartHeight,
       groupBarsAreaHeight,
       groupHeight,
-      tallestXAxisLabel,
     };
   }, [
-    bandwidth,
     chartDimensions,
     isSimple,
     isStacked,
     seriesLength,
     singleBarCount,
-    tallestXAxisLabel,
+    labelHeight,
   ]);
 }